extract to a hook
